refactor: refactored product entity and deleted product variant
